Add training tuner (ray.tune)
A new script tune_train.py serves the purpose of meta-training, where
arbitrary hyperparameters are tuned using ray.tune.

Other unrelated changes:
* Priority dtype is changed to float64 (from float32).
* Remove make_game() in favor of Game ctor.
* Add preprocess_state() game callback (unused).
* Others minor.

+++ d3q/apps/tune_replaymemory.py
@@ -40,7 +40,7 @@ def benchmark_replaymemory(config):
                           max_branch_bucket_capacity=config['max_branch_bucket_capacity'])
 
         batch = np.arange(0, BATCH_SIZE, 1, dtype=np.int32)
-        priorities = np.full(shape=(BATCH_SIZE,), fill_value=1.0, dtype=np.float32)
+        priorities = np.full(shape=(BATCH_SIZE,), fill_value=1.0, dtype=np.float64)
 
         start_time = time.time()
         for step in range(NUM_STEPS):
