+++ d3q/tests/test_replaymemory.py
@@ -27,39 +27,39 @@ def test_priority_tree_1():
     # | 0 0 0 0 | 0 0 0 0 | 0 0 0 |
     assert root.priority_sum == pytest.approx(0.0)
 
-    root.update_priority_range(0, 11, np.full((11,), 1.0, dtype=np.float32))
+    root.update_priority_range(0, 11, np.full((11,), 1.0, dtype=np.float64))
     # | 1 1 1 1 | 1 1 1 1 | 1 1 1 |
     assert root.priority_sum == pytest.approx(11.0)
 
-    root.update_priority_range(1, 3, np.full((2,), 2.0, dtype=np.float32))
+    root.update_priority_range(1, 3, np.full((2,), 2.0, dtype=np.float64))
     # | 1 2 2 1 | 1 1 1 1 | 1 1 1 |
     assert root.priority_sum == pytest.approx(13.0)
 
-    root.update_priority_range(2, 10, np.full((8,), 3.0, dtype=np.float32))
+    root.update_priority_range(2, 10, np.full((8,), 3.0, dtype=np.float64))
     # | 1 2 3 3 | 3 3 3 3 | 3 3 1 |
     assert root.priority_sum == pytest.approx(28.0)
 
-    root.update_priority_range(3, 6, np.full((3,), 4.0, dtype=np.float32))
+    root.update_priority_range(3, 6, np.full((3,), 4.0, dtype=np.float64))
     # | 1 2 3 4 | 4 4 3 3 | 3 3 1 |
     assert root.priority_sum == pytest.approx(31.0)
 
-    root.update_priority_range(6, 9, np.full((3,), 4.0, dtype=np.float32))
+    root.update_priority_range(6, 9, np.full((3,), 4.0, dtype=np.float64))
     # | 1 2 3 4 | 4 4 4 4 | 4 3 1 |
     assert root.priority_sum == pytest.approx(34.0)
 
-    root.update_priority_range(1, 2, np.full((1,), 5.0, dtype=np.float32))
+    root.update_priority_range(1, 2, np.full((1,), 5.0, dtype=np.float64))
     # | 1 5 3 4 | 4 4 4 4 | 4 3 1 |
     assert root.priority_sum == pytest.approx(37.0)
 
-    root.update_priority_range(5, 7, np.full((2,), 6.0, dtype=np.float32))
+    root.update_priority_range(5, 7, np.full((2,), 6.0, dtype=np.float64))
     # | 1 5 3 4 | 4 6 6 4 | 4 3 1 |
     assert root.priority_sum == pytest.approx(41.0)
 
-    root.update_priority_range(0, 11, np.full((11,), 1.0, dtype=np.float32))
+    root.update_priority_range(0, 11, np.full((11,), 1.0, dtype=np.float64))
     # | 1 1 1 1 | 1 1 1 1 | 1 1 1 |
     assert root.priority_sum == pytest.approx(11.0)
 
-    root.update_priority_range(0, 11, np.full((11,), 0.0, dtype=np.float32))
+    root.update_priority_range(0, 11, np.full((11,), 0.0, dtype=np.float64))
     # | 0 0 0 0 | 0 0 0 0 | 0 0 0 |
     assert root.priority_sum == pytest.approx(0.0)
 
@@ -73,15 +73,15 @@ def test_priority_tree_2():
         root = rm.root_bucket
         assert root.priority_sum == pytest.approx(0.0)
 
-        root.update_priority_range(0, 10001, np.full((10001,), 1.0, dtype=np.float32))
+        root.update_priority_range(0, 10001, np.full((10001,), 1.0, dtype=np.float64))
         assert root.priority_sum == pytest.approx(10001.0)
 
-        root.update_priority_range(0, 10001, np.full((10001,), 2.0, dtype=np.float32))
+        root.update_priority_range(0, 10001, np.full((10001,), 2.0, dtype=np.float64))
         assert root.priority_sum == pytest.approx(20002.0)
 
         start_index = 1234
         len = 4567
-        root.update_priority_range(start_index, start_index+len, np.full((len,), 4.0, dtype=np.float32))
+        root.update_priority_range(start_index, start_index+len, np.full((len,), 4.0, dtype=np.float64))
         assert root.priority_sum == pytest.approx(20002.0 + 2.0*len)
 
 
@@ -93,7 +93,7 @@ def test_priority_tree_3():
     root = rm.root_bucket
     assert root.priority_sum == pytest.approx(0.0)
 
-    t = np.cumsum(np.full((1000,), 0.01, dtype=np.float32))
+    t = np.cumsum(np.full((1000,), 0.01, dtype=np.float64))
     s = np.sum(t)
 
     root.update_priority_range(0, 1000, t)
@@ -107,7 +107,7 @@ def test_sample_1():
                       max_branch_bucket_capacity=4)
     for i in range(64):
         eb = np.array([i], dtype=np.int32)
-        rm.memorize(eb, priorities=np.array([10.0*i + 0.1], dtype=np.float32))
+        rm.memorize(eb, priorities=np.array([10.0*i + 0.1], dtype=np.float64))
 
     rm.sample_random(128)
 
@@ -119,9 +119,9 @@ def test_sample_2():
                       max_branch_bucket_capacity=2)
 
     eb = np.arange(0, 10, 1, dtype=np.int32)
-    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float32))
+    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float64))
 
-    sp = np.arange(0.0, 10.0, 0.01, dtype=np.float32)
+    sp = np.arange(0.0, 10.0, 0.01, dtype=np.float64)
     recs, virt_indices = rm.sample_from_points(sp, sample_points_sorted=True)
 
     assert (np.unique(recs) == eb).all()
@@ -133,25 +133,25 @@ def test_update_prio_1():
                       max_leaf_bucket_capacity=4,
                       max_branch_bucket_capacity=3)
     eb = np.arange(0, 14, 1, dtype=np.int32)
-    rm.memorize(eb, priorities=np.full(shape=(14,), fill_value=1.0, dtype=np.float32))
+    rm.memorize(eb, priorities=np.full(shape=(14,), fill_value=1.0, dtype=np.float64))
     # | 1 1 1 1 1 | 1 1 1 1 1 | 1 1 1 1 |
     assert rm.root_bucket.priority_sum == pytest.approx(14.0)
 
-    sp = np.arange(0.5, 14, 1.0, dtype=np.float32)
+    sp = np.arange(0.5, 14, 1.0, dtype=np.float64)
     sp = np.random.permutation(sp)
     recs, virt_indices = rm.sample_from_points(sp)
     assert (recs == eb).all()
     assert (virt_indices == eb).all()
 
-    rm.update_priorities(np.array([1, 3, 5, 7, 9, 11, 13], dtype=np.int64), np.full((7,), 2.0, dtype=np.float32))
+    rm.update_priorities(np.array([1, 3, 5, 7, 9, 11, 13], dtype=np.int64), np.full((7,), 2.0, dtype=np.float64))
     # | 1 2 1 2 1 | 2 1 2 1 2 | 1 2 1 2 |
     assert rm.root_bucket.priority_sum == pytest.approx(21.0)
 
-    rm.update_priorities(np.array([2, 6, 12], dtype=np.int64), np.full((3,), 3.0, dtype=np.float32))
+    rm.update_priorities(np.array([2, 6, 12], dtype=np.int64), np.full((3,), 3.0, dtype=np.float64))
     # | 1 2 3 2 1 | 2 3 2 1 2 | 1 2 3 2 |
     assert rm.root_bucket.priority_sum == pytest.approx(27.0)
 
-    rm.update_priorities(np.array([0, 10], dtype=np.int64), np.full((2,), 0.0, dtype=np.float32))
+    rm.update_priorities(np.array([0, 10], dtype=np.int64), np.full((2,), 0.0, dtype=np.float64))
     # | 0 2 3 2 1 | 2 3 2 1 2 | 0 2 3 2 |
     assert rm.root_bucket.priority_sum == pytest.approx(25.0)
 
@@ -162,23 +162,23 @@ def test_update_prio_2():
                       max_leaf_bucket_capacity=10,
                       max_branch_bucket_capacity=1)
     eb = np.arange(0, 10, 1, dtype=np.int32)
-    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float32))
-    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float32))
+    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float64))
+    rm.memorize(eb, priorities=np.full(shape=(10,), fill_value=1.0, dtype=np.float64))
     # | 1 1 1 1 1 1 1 1 1 1 |
     assert rm.root_bucket.priority_sum == pytest.approx(10.0)
 
-    sp = np.arange(0.5, 10, 1.0, dtype=np.float32)
+    sp = np.arange(0.5, 10, 1.0, dtype=np.float64)
     sp = np.random.permutation(sp)
     recs, virt_indices = rm.sample_from_points(sp)
     assert (recs == eb).all()
     assert (virt_indices == (eb + 10)).all()
 
     eb = np.arange(10, 14, 1, dtype=np.int32)
-    rm.memorize(eb, priorities=np.full(shape=(4,), fill_value=2.0, dtype=np.float32))
+    rm.memorize(eb, priorities=np.full(shape=(4,), fill_value=2.0, dtype=np.float64))
     # | 2 2 2 2 1 1 1 1 1 1 |
     assert rm.root_bucket.priority_sum == pytest.approx(14.0)
 
-    sp = np.arange(0.0, rm.root_bucket.priority_sum, 0.07, dtype=np.float32)
+    sp = np.arange(0.0, rm.root_bucket.priority_sum, 0.07, dtype=np.float64)
     recs, virt_indices = rm.sample_from_points(sp, sample_points_sorted=True)
 
     assert min(recs) == 4
@@ -187,10 +187,10 @@ def test_update_prio_2():
     assert max(virt_indices) == 23
 
     eb = np.arange(14, 17, 1, dtype=np.int32)
-    rm.memorize(eb, priorities=np.full(shape=(3,), fill_value=3.0, dtype=np.float32))
+    rm.memorize(eb, priorities=np.full(shape=(3,), fill_value=3.0, dtype=np.float64))
     # | 2 2 2 2 3 3 3 1 1 1 |
     assert rm.root_bucket.priority_sum == pytest.approx(20.0)
 
-    rm.update_priorities(virt_indices, priorities=np.full(virt_indices.shape, fill_value=4.0, dtype=np.float32))
+    rm.update_priorities(virt_indices, priorities=np.full(virt_indices.shape, fill_value=4.0, dtype=np.float64))
     # | 4 4 4 4 3 3 3 4 4 4 |
     assert rm.root_bucket.priority_sum == pytest.approx(37.0)
